+++ src/game/mario_actions_airborne.c
@@ -17,7 +17,7 @@
 
 #include "config.h"
 
-extern u32 gGravityMode;
+extern u8 gGravityMode;
 
 void play_flip_sounds(struct MarioState *m, s16 frame1, s16 frame2, s16 frame3) {
     s32 animFrame = m->marioObj->header.gfx.animInfo.animFrame;
